mcp-platform-demo/mcp-gateway/gateway.py
import asyncio
import getpass
import json
import os
import time
from datetime import datetime
from typing import Any, Dict

import yaml

os.getlogin = getpass.getuser

import docker
import redis
import websockets
from mcp.server.fastmcp import FastMCP

# Load configuration
with open("config.yaml", "r") as f:
    config = yaml.safe_load(f)

# Initialize MCP server
mcp = FastMCP(config["server"]["name"])

# Global variables
websocket_clients = set()
docker_client = docker.from_env()
operation_history = []
redis_client = redis.Redis(
    host=config["services"]["redis"]["host"], port=config["services"]["redis"]["port"]
)


@mcp.tool()
def create_api(
    name: str,
    description: str = "",
    type: str = "rest",
    authentication: bool = True,
    rate_limit: int = 1000,
) -> str:
    """Create a new API with complete governance, observability, and security.

    Args:
        name: API name
        description: API description
        type: API type (rest, graphql, grpc)
        authentication: Enable authentication
        rate_limit: Requests per minute limit

    Returns:
        Creation report with details and next steps
    """
    start_time = time.time()

    # Record operation
    operation = {
        "id": f"op_{int(time.time() * 1000)}",
        "tool": "create_api",
        "arguments": {
            "name": name,
            "description": description,
            "type": type,
            "authentication": authentication,
            "rate_limit": rate_limit,
        },
        "timestamp": datetime.now().isoformat(),
        "status": "started",
    }
    operation_history.append(operation)

    # Broadcast to dashboard
    asyncio.create_task(
        broadcast_event({"type": "operation_start", "operation": operation})
    )

    # Mock implementation steps
    steps = [
        "📝 Analyzing requirements",
        "🔧 Generating OpenAPI specification",
        "🌐 Configuring in Kong Gateway",
        "🔒 Setting up authentication",
        "⚡ Applying rate limiting",
        "📊 Creating Grafana dashboard",
        "🔍 Configuring Jaeger tracing",
        "🚀 Making initial deployment",
        "✅ Running integration tests",
        "📚 Generating documentation",
    ]

    results = []
    for step in steps:
        # Broadcast step progress
        asyncio.create_task(
            broadcast_event(
                {"type": "step_progress", "step": step, "status": "in_progress"}
            )
        )

        # Simulate step execution (would contain actual implementation)
        time.sleep(1)

        # Broadcast step completion
        asyncio.create_task(
            broadcast_event(
                {"type": "step_progress", "step": step, "status": "completed"}
            )
        )

        results.append(f"✅ {step}")

    # Mock metrics
    metrics = {
        "status": "🟢 Healthy",
        "rps": 127,
        "p99_latency": 45,
        "error_rate": 0.1,
        "uptime": 100,
    }

    # Update operation status
    operation["status"] = "completed"
    operation["duration"] = time.time() - start_time

    # Broadcast completion
    asyncio.create_task(
        broadcast_event({"type": "operation_complete", "operation": operation})
    )

    return f"""# 🚀 API Created Successfully!

## API: {name}
**Description**: {description}
**Type**: {type.upper()}

### Creation Summary:
{chr(10).join(results)}

### Access Points:
- **Gateway URL**: http://localhost:8000/api/{name}
- **Swagger UI**: http://localhost:8082/{name}
- **Grafana Dashboard**: http://localhost:3000/d/{name}
- **Jaeger Tracing**: http://localhost:16686/trace/{name}

### Applied Governance:
- ✅ Rate Limiting: {rate_limit} req/min
- ✅ JWT Authentication: {"Enabled" if authentication else "Disabled"}
- ✅ Request/Response Validation
- ✅ CORS Configured
- ✅ Circuit Breaker Active
- ✅ Retry Policy: 3 attempts

### Real-time Metrics:
```
Status: {metrics["status"]}
Requests/sec: {metrics["rps"]}
P99 Latency: {metrics["p99_latency"]}ms
Error Rate: {metrics["error_rate"]}%
Uptime: {metrics["uptime"]}%
```

### Next Steps:
1. Test your API: `curl http://localhost:8000/api/{name}/health`
2. Explore documentation in Swagger UI
3. Configure webhooks if needed
4. Monitor metrics in Grafana

### Example Code:
```python
import requests

# Example API call
headers = {{
    'Authorization': 'Bearer your-token-here',
    'Content-Type': 'application/json'
}}

response = requests.get(
    'http://localhost:8000/api/{name}/endpoint',
    headers=headers
)

print(response.json())
```

The API is ready for use and being monitored! 🎉
"""


@mcp.tool()
def investigate_problem(
    description: str, service: str = "all", auto_fix: bool = False
) -> str:
    """Investigate and diagnose platform problems.

    Args:
        description: Problem description
        service: Target service (or "all")
        auto_fix: Enable automatic problem fixing

    Returns:
        Investigation report with root cause and solution
    """
    start_time = time.time()

    # Record operation
    operation = {
        "id": f"op_{int(time.time() * 1000)}",
        "tool": "investigate_problem",
        "arguments": {
            "description": description,
            "service": service,
            "auto_fix": auto_fix,
        },
        "timestamp": datetime.now().isoformat(),
        "status": "started",
    }
    operation_history.append(operation)

    # Broadcast to dashboard
    asyncio.create_task(
        broadcast_event({"type": "operation_start", "operation": operation})
    )

    # Investigation steps
    steps = [
        "🔍 Analyzing error logs",
        "📊 Checking anomalous metrics",
        "🔗 Examining dependencies",
        "🚀 Reviewing recent deployments",
        "💻 Analyzing resource usage",
        "🌐 Checking network latency",
        "🔍 Correlating events",
        "🧠 Root cause analysis",
    ]

    findings = []
    for step in steps:
        # Broadcast step progress
        asyncio.create_task(
            broadcast_event(
                {"type": "investigation_step", "step": step, "status": "analyzing"}
            )
        )

        # Simulate step execution
        time.sleep(1)

        # Add mock finding
        finding = {
            "step": step,
            "summary": f"Analysis completed for {step}",
            "details": f"Detailed analysis for {service} regarding {description}",
        }
        findings.append(finding)

        # Broadcast step completion
        asyncio.create_task(
            broadcast_event(
                {
                    "type": "investigation_step",
                    "step": step,
                    "status": "completed",
                    "finding": finding,
                }
            )
        )

    # Mock root cause (based on service and description)
    if "latency" in description.lower():
        root_cause = {
            "description": "Database connection pool exhaustion causing request queuing",
            "affected_services": [service]
            if service != "all"
            else ["payment-service", "user-service"],
            "error_increase": "+150%",
            "latency_impact": "+300ms p99",
            "users_affected": "~40%",
        }
    elif "error" in description.lower():
        root_cause = {
            "description": "Invalid configuration after recent deployment causing 500 errors",
            "affected_services": [service] if service != "all" else ["order-service"],
            "error_increase": "+80%",
            "latency_impact": "+20ms p99",
            "users_affected": "~15%",
        }
    else:
        root_cause = {
            "description": "Memory leak in service causing gradual performance degradation",
            "affected_services": [service] if service != "all" else ["user-service"],
            "error_increase": "+5%",
            "latency_impact": "+100ms p99",
            "users_affected": "~25%",
        }

    # Mock solution
    solution = {
        "steps": """1. Increase database connection pool size
2. Add connection timeout handling
3. Implement retry logic with exponential backoff
4. Add Redis caching layer for frequent queries
5. Update monitoring thresholds""",
        "commands": """# Restart affected service with new configuration
kubectl rollout restart deployment payment-service

# Apply database optimization
kubectl exec -it postgres-0 -- psql -U postgres -c "ALTER SYSTEM SET max_connections = 200;"
kubectl exec -it postgres-0 -- psql -U postgres -c "SELECT pg_reload_conf();"

# Verify fix
kubectl logs -f deployment/payment-service""",
        "prevention": """1. Add connection pool metrics to dashboards
2. Create alert for connection usage > 80%
3. Implement circuit breaker pattern
4. Add database connection metrics to pre-deployment checks
5. Schedule regular database optimization""",
    }

    # Update operation status
    operation["status"] = "completed"
    operation["duration"] = time.time() - start_time

    # Broadcast completion
    asyncio.create_task(
        broadcast_event({"type": "operation_complete", "operation": operation})
    )

    # Generate findings text
    findings_text = "\n".join([f"🔍 {f['step']}: {f['summary']}" for f in findings])

    return f"""# 🔍 Investigation Report

## Problem: {description}
**Service**: {service}
**Time**: {datetime.now().strftime("%Y-%m-%d %H:%M:%S")}

### Investigation Timeline:
{findings_text}

### Root Cause Analysis:
**{root_cause["description"]}**

### Detected Impact:
- 🔴 Affected Services: {", ".join(root_cause["affected_services"])}
- 📊 Error Rate: {root_cause["error_increase"]}
- ⏱️ Latency Impact: {root_cause["latency_impact"]}
- 👥 Affected Users: {root_cause["users_affected"]}

### Recommended Solution:
{solution["steps"]}

### Immediate Actions:
```bash
{solution["commands"]}
```

### Visualizations:
- [Incident Dashboard](http://localhost:3000/d/incident-{int(time.time())})
- [Problem Traces](http://localhost:16686/trace/{service})
- [Detailed Logs](http://localhost:5601/app/logs?query={description})

### Future Prevention:
{solution["prevention"]}

### Remediation Status:
{"✅ Automatic fix applied successfully!" if auto_fix else "⏳ Waiting for manual solution application"}

### Post-Incident Monitoring:
- Alert configured to detect recurrence
- Specific dashboard created
- Runbook updated with this scenario
"""


# WebSocket handling
async def broadcast_event(event: Dict[str, Any]):
    """Send events to all WebSocket clients"""
    if websocket_clients:
        message = json.dumps(event)
        disconnected = set()

        for client in websocket_clients:
            try:
                await client.send(message)
            except websockets.exceptions.ConnectionClosed:
                disconnected.add(client)

        # Remove disconnected clients
        websocket_clients.difference_update(disconnected)


async def websocket_handler(websocket, path):
    """Handle WebSocket connections"""
    websocket_clients.add(websocket)
    try:
        await websocket.wait_closed()
    finally:
        websocket_clients.remove(websocket)


# Start WebSocket server
async def start_websocket_server():
    """Start WebSocket server for dashboard communication"""
    ws_port = config["server"]["websocket_port"]
    print(f"Starting WebSocket server on port {ws_port}")

    await websockets.serve(websocket_handler, "0.0.0.0", ws_port)


# Main function
if __name__ == "__main__":
    # Start WebSocket server
    asyncio.run(start_websocket_server())
