mcp-platform-demo/mcp-gateway/requirements.txt
mcp>=1.7.0
fastapi>=0.95.1
uvicorn>=0.22.0
pydantic>=2.0.0
websockets>=11.0.3
docker>=6.1.3
requests>=2.31.0
aiohttp>=3.8.5
redis>=4.6.0
pyyaml>=6.0.1
jinja2>=3.1.2
prometheus-client>=0.17.1
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
asyncio>=3.4.3
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0 